Finished the countdown clock display
- Reconfigured the Chrono class to use some common conversion factors
- Changed some of the time values in Event to doubles from integers
- Wrote all of the GUI logic for clock display: digit drawing methods, unicode relationships, etc

On branch develop

Changes to be committed:
	modified:   chrono.cpp
	modified:   chrono.hpp
	modified:   event.hpp
	modified:   gui.cpp
	modified:   panel.cpp
	modified:   panel.hpp

+++ src/event.hpp
@@ -37,7 +37,7 @@ class CountdownTimer : public GameEvent {
 		CountdownTimer(uint remainingTime);
 		void update(); // add or remove time from the timer
 		void execute();
-		int getRemainingTime() { return remainder; }
+		double getRemainingTime() { return remainder; }
 		void advanceTime(double increment);
 		std::string getRemainingTimeString();
